// src/app/models/referencias.ts
export class Referencias {
    numIdReferencia: number;
    strDominio: string;
    strRangoValor: string;
    strDescripcion: string;
    strValorParametro: number;
    strEstado: string;
    strTipoProductoFk: string;
    strTipoClaveFk: string;
    strTipoCuentaFk: string;
    strTipoIdFk: string;
    strEsActivoFk: string;
}